requirements.txt

numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.5.0
pandas>=1.3.0
h5py>=3.0.0
pyyaml>=6.0
pathlib2>=2.3.0 

torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
tqdm>=4.64.0

